ENH: S'ha passat de double a quint64 perquè no hi ha decimals.

+++ starviewer/src/inputoutput/qcreatedicomdir.cpp
@@ -51,7 +51,7 @@ QCreateDicomdir::QCreateDicomdir(QWidget *parent)
     createConnections();
 
     //per defecte gravem al disc dur per tant, l'espai és il·limitat
-    m_DiskSpace = ( double ) 9999999 * (double) ( 1024 * 1024 );
+    m_DiskSpace = ( quint64 ) 9999999 * ( quint64 ) ( 1024 * 1024 );
 
     setWidthColumns();
 }
@@ -156,7 +156,7 @@ void QCreateDicomdir::addStudy( DICOMStudy study )
 {
     CacheImageDAL cacheImageDAL;
     DicomMask imageMask;
-    double studySize;
+    quint64 studySize;
     Status state;
 
     if ( !studyExists( study.getStudyUID() ) )
@@ -416,7 +416,7 @@ void QCreateDicomdir::removeSelectedStudy()
     DicomMask imageMask;
     CacheImageDAL cacheImageDAL;
     Status state;
-    double studySize;
+    quint64 studySize;
     QList<QTreeWidgetItem *> selectedStudies;
 
     selectedStudies = m_dicomdirStudiesList->selectedItems();
@@ -632,14 +632,14 @@ void QCreateDicomdir::deviceChanged( int index )
         case HardDisk:
             m_stackedWidget->setCurrentIndex(1);
             // per gravar al disc no hi ha màxim TODO això no es del tot cert, caldria comprovar l'espai de disc
-            m_DiskSpace = ( double ) 9999999 * ( double ) ( 1024 * 1024 );
+            m_DiskSpace = ( quint64 ) 9999999 * ( quint64 ) ( 1024 * 1024 );
             break;
         case CDROM:
                 m_stackedWidget->setCurrentIndex(0);
                 if( sizeInMB < 700 )
                 {
                     m_progressBarOcupat->setMaximum( 700 );
-                    m_DiskSpace = ( double ) 700 * ( double ) ( 1024 * 1024 ); // convertim a bytes capacaticat cd
+                    m_DiskSpace = ( quint64 ) 700 * ( quint64 ) ( 1024 * 1024 ); // convertim a bytes capacaticat cd
                     m_progressBarOcupat->repaint();
                 }
                 else
@@ -652,7 +652,7 @@ void QCreateDicomdir::deviceChanged( int index )
                 if( sizeInMB < 4400 )
                 {
                     m_progressBarOcupat->setMaximum( 4400 );
-                    m_DiskSpace = ( double ) 4400 * ( double ) ( 1024 * 1024 ); //convertim a bytes capacitat dvd
+                    m_DiskSpace = ( quint64 ) 4400 * ( quint64 ) ( 1024 * 1024 ); //convertim a bytes capacitat dvd
                     m_progressBarOcupat->repaint();
                 }
                 else
